+++ src/main/java/com/netcracker/smarthome/business/chart/configuration/MultiChartConfig.java
@@ -14,7 +14,7 @@ public class MultiChartConfig extends DefaultChartConfig implements ChartConfigu
     private String title;
     private ChartService chartService;
 
-    public MultiChartConfig(SmartHome smartHome, long chartId, String chartTitle, ChartService chartService, double refreshInterval, String chartType, String chartInterval) {
+    public MultiChartConfig(SmartHome smartHome, long chartId, String chartTitle, ChartService chartService, long refreshInterval, String chartType, String chartInterval) {
         super(smartHome, chartId, refreshInterval, chartType, chartInterval);
         this.title = chartTitle;
         this.chartService = chartService;
